Fix ACE module files failing to load
When the ACE editor is initialized, by default the base path used to
load module files is set to the directory of the "ace.js" file. Since
the JavaScript files were merged "ace.js" is no longer added to the page
as a single script file, so the base path was no longer set, and thus
the module files failed to load. Now the right base path is explicitly
set in the editor configuration.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>