Add plain text previews
The default plain text previews always assume that the text file uses
UTF-8 encoding, so when it uses a different one the characters are not
properly parsed by the browser. The "public" endpoint from this app
tries to guess the actual encoding of the file and then converts the
data sent to the browser to UTF-8. Thus, now the default plain text
previews are overriden with the preview generated by this app.

Note that this does not prevent the default plain text preview from
being generated, though; it is simply hidden. Following the same
approach used for Markdown previews, ".preview" is added to the
"#imgframe" element, which is the parent element for the default plain
text preview. This removes the default plain text preview thanks to the
CSS rules. However, as the "overrider" plain text preview is also a
descendant of ".preview", a special CSS class has to be added to the
"overrider" plain text preview to prevent it from being hidden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>