Do not keep FileInfoModels returned by "getModelForFile"
FileInfoModels returned by "getModelForFile" are just temporary, and
they should be used and forgotten instead of being kept by apps. It is
not guaranteed that different calls to "getModelForFile" for the same
file returns the same FileInfoModel object, so changes in one model
instance could be not known by other model instances (and their views),
and this could lead to different bugs (for example, a file list entry
being duplicated when a text file is edited or favourited after being
created).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>